New season & Prestige restore order
Wait for the prestige check to finish before resetting the season, if it executes in the same time, there is a chance the player could get his prestige restored after the season reset, and can maintain his prestige, so admins need to manually restore.

This edge case already happened plenty of times.